Conflicts:
	pom.xml
	src/main/db/changelog/changelog.xml
	src/main/java/cn/explink/controller/ApplyEditDeliverystateController.java
	src/main/java/cn/explink/controller/CwbOrderController.java
	src/main/java/cn/explink/dao/BranchDAO.java
	src/main/java/cn/explink/domain/Branch.java
	src/main/java/cn/explink/service/BranchService.java
	src/main/webapp/WEB-INF/jsp/branch/edit.jsp

+++ src/main/java/cn/explink/enumutil/IsmpsflagEnum.java
@@ -3,15 +3,15 @@ package cn.explink.enumutil;
 public enum IsmpsflagEnum {
 	no(0, "普通件"), yes(1, "一票多件"), ;
 
-	private long value;
+	private int value;
 	private String text;
 
-	private IsmpsflagEnum(long value, String text) {
+	private IsmpsflagEnum(int value, String text) {
 		this.value = value;
 		this.text = text;
 	}
 
-	public long getValue() {
+	public int getValue() {
 		return this.value;
 	}
